Pravljenje Gate autorizacije

+++ app/Http/Controllers/QuestionsCotroller.php
@@ -66,6 +66,9 @@ class QuestionsCotroller extends Controller
      */
     public function edit(Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Ne dozvoljen pristup");
+        }
         return view('questions.edit', compact('question'));
     }
 
@@ -78,6 +81,10 @@ class QuestionsCotroller extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (\Gate::denies('update-question', $question)) {
+            abort(403, "Ne dozvoljen pristup");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', "Tvoje pitanje je azurirano");
@@ -91,6 +98,9 @@ class QuestionsCotroller extends Controller
      */
     public function destroy(Question $question)
     {
+        if (\Gate::denies('delete-question', $question)) {
+            abort(403, "Ne dozvoljen pristup");
+        }
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', "Tvoje pitanje je uspesno obrisano");
